+++ examples/advanced/color/bezierEasing.less
