+++ examples/advanced/color/colorPalette.less
