+++ examples/advanced/color/tinyColor.less
