+++ examples/import.less
@@ -1,4 +1,3 @@
-@import "javascript.less";
 .foo {
   background: #900;
   border: 1px solid @color;
